deploy_aic_lstm-main/requirements.txt
fastapi
keras
pillow
python-multipart
opencv-python-headless
uvicorn
gunicorn==19.9.0
Jinja2
numpy
scikit-learn